requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
build==1.3.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
chromadb==1.3.5
click==8.3.1
coloredlogs==15.0.1
colorlog==6.10.1
dataclasses-json==0.6.7
dill==0.4.0
distro==1.9.0
docling==2.64.1
docling-core==2.54.1
docling-ibm-models==3.10.3
docling-parse==4.7.2
durationpy==0.10
et-xmlfile==2.0.0
faker==38.2.0
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
gitpython==3.1.45
google-ai-generativelanguage==0.6.18
google-api-core==2.28.1
google-auth==2.43.0
google-genai==1.54.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib-metadata==8.7.0
importlib-resources==6.5.2
jinja2==3.1.6
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
langchain==0.3.12
langchain-chroma==0.2.4
langchain-community==0.3.12
langchain-core==0.3.63
langchain-google-genai==2.1.5
langchain-text-splitters==0.3.8
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.15
langsmith==0.2.11
latex2mathml==3.78.1
loguru==0.7.3
lxml==6.0.2
markdown-it-py==4.0.0
marko==2.2.1
markupsafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.2.0
mpire==2.10.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
mypy-extensions==1.1.0
narwhals==2.13.0
networkx==3.6.1
numpy==2.3.5
oauthlib==3.3.1
ocrmac==1.0.0
omegaconf==2.3.0
onnxruntime==1.23.2
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.39.0
opentelemetry-exporter-otlp-proto-common==1.39.0
opentelemetry-exporter-otlp-proto-grpc==1.39.0
opentelemetry-proto==1.39.0
opentelemetry-sdk==1.39.0
opentelemetry-semantic-conventions==0.60b0
orjson==3.11.5
ormsgpack==1.12.0
overrides==7.7.0
packaging==24.2
pandas==2.3.3
pillow==11.3.0
pluggy==1.6.0
polyfactory==3.1.0
posthog==5.4.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.2
psutil==7.1.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybase64==1.4.3
pyclipper==1.4.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pydeck==0.9.1
pygments==2.19.2
pylatexenc==2.10
pyobjc-core==12.1
pyobjc-framework-cocoa==12.1
pyobjc-framework-coreml==12.1
pyobjc-framework-quartz==12.1
pyobjc-framework-vision==12.1
pypdfium2==4.30.0
pypika==0.48.9
pyproject-hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-pptx==1.0.2
pytz==2025.2
pyyaml==6.0.3
rapidocr==3.4.3
referencing==0.37.0
regex==2025.11.3
reportlab==4.4.5
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
rtree==1.4.1
safetensors==0.7.0
scipy==1.16.3
semchunk==2.2.2
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
soupsieve==2.8
sqlalchemy==2.0.45
streamlit==1.52.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.3
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-java==0.23.5
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
typer==0.19.2
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.3.0
uuid-utils==0.12.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
